+++ src/service/worker/runtime/http/middlewares/cancellationToken.ts
@@ -8,7 +8,8 @@ export async function cancellationToken<
   U extends RecorderState,
   V extends ParamsContext
 >(ctx: ServiceContext<T, U, V>, next: () => Promise<void>) {
-  if (ctx.method.toUpperCase() === ('GET' || 'OPTIONS' || 'HEAD')) {
+  const cancellableMethods = new Set(['GET', 'OPTIONS', 'HEAD'])
+  if (cancellableMethods.has(ctx.method.toUpperCase())) {
     ctx.vtex.cancellation = {
       cancelable: true,
       cancelled: false,